move BankCreationFreezingProgress to its own file
